+++ lib/components/narrative/metro/metro-itinerary.tsx
@@ -27,8 +27,8 @@ import ItineraryBody from '../line-itin/connected-itinerary-body'
 import NarrativeItinerary from '../narrative-itinerary'
 import SimpleRealtimeAnnotation from '../simple-realtime-annotation'
 
+import { DepartureTimesList } from './departure-times-list'
 import {
-  DepartureTimesList,
   getFirstTransitLegStop,
   getFlexAttirbutes,
   getItineraryRoutes,
